Refactor code to allow executing a handler both on syscall exit and enter.
Currently, the register values are only scraped on syscall exit.
Unfortunately, there are cases, when the input arguments are not
preserved, e.g. while executing a Python script with a shabang, for
example ./script.py

Now we support both types of handlers.

Besides, we took advantage of thread_local! to clean up the random byte
generation.